refactor: Split tests into separate files for better organization
